refactor(cli,web): kill duplicate utilities + introduce shared types
## CLI — 16 × `function truncate` → 1 shared import

16 command modules had their own copy of a 5-8-line `truncate(str, len)`
helper (112+ matching lines across `cli/lib/commands/`). Three distinct
flavors had drifted out of step:

  - 6 simple cut-at-char: diagnostics, runtime, inventory, agent,
    federation, traceability  →  now import `truncate` from `lib/ui/ansi.mjs`
  - 10 word-boundary-aware:  contributors, providers, lineage, assets,
    specs, news, friction, governance, services, ideas
    →  now import `{ truncateWords as truncate }` so existing call sites
       don't have to change
  - 1 defensive wrapper (ops.mjs): left as-is because it also coerces
    non-string input via `String(value || "").trim()` before truncating —
    not safe to swap for the shared helper without auditing every caller

Added `truncateWords(str, max)` to `cli/lib/ui/ansi.mjs` alongside the
existing `truncate(str, max=60)`. The new helper trims to `max` chars
then walks back to the last space, falling through to a hard cut if the
first word alone is longer than 40% of `max`. This matches the
historical behavior the 10 tables depended on.

Verification: `node cli/bin/cc.mjs ideas 3` against prod API still
returns the curated-super-idea table with correctly-truncated columns;
`import()` smoke tests pass for ideas/specs/agent/governance modules.

## Web — shared types module

Eight files had their own redefinitions of `IdeaWithScore` (6 copies)
and `IdeaQuestion` (6 copies). The shapes had drifted — some included
`actual_cost`, some didn't; some included `interfaces`, some didn't;
some included the `is_curated` / `pillar` fractal-restructuring fields,
some didn't. The risk was that a page would read a field the API
doesn't return (runtime `undefined`) or, worse, fail to handle a field
it DOES return.

New module: `web/lib/types/ideas.ts` exports a single `IdeaWithScore`
+ `IdeaQuestion` aligned with `api/app/models/idea.py`. Fields the API
always returns are required (`actual_cost`, `open_questions`,
`interfaces`). Fields present only on richer views (`stage`,
`resistance_risk`, `idea_type`, hierarchy, `is_curated`, `pillar`)
stay optional — pages that need a narrower contract can use
`Pick<IdeaWithScore, ...>` instead of redeclaring the whole type.

Re-export barrel at `web/lib/types/index.ts` (pages import from
`@/lib/types`).

Migrated (local types deleted, shared types imported):
  - web/app/ideas/page.tsx
  - web/app/ideas/[idea_id]/page.tsx
  - web/app/page.tsx
  - web/app/invest/page.tsx
  - web/app/demo/page.tsx
  - web/app/contribute/page.tsx
  - web/components/ideas/IdeaDetailTabs.tsx
  - web/components/ideas/IdeasListView.tsx

Verification: `tsc --noEmit --skipLibCheck` exits 0. No runtime changes.

## Notes

- This is pure DRY cleanup. Zero new features, zero behavior changes,
  zero breaking changes. Production response shapes are unchanged.
- The CLI-side `ops.mjs` truncate is intentionally left alone — future
  PR should either add a `coerce` overload to the shared helper or
  audit ops callers and inline the coercion.
- Follow-up PRs in this series will add more shared types
  (`SpecRegistryEntry`, `Contributor`, `FlowItem`, etc.) as the
  refactoring opportunities line up.
