+++ cli/lib/commands/inventory.mjs
@@ -15,12 +15,8 @@
  */
 
 import { get, post } from "../api.mjs";
+import { truncate } from "../ui/ansi.mjs";
 
-function truncate(str, len) {
-  if (!str) return "";
-  if (str.length <= len) return str;
-  return str.slice(0, len - 3) + "...";
-}
 
 function fmt(val, pad = 0) {
   const s = val == null ? "—" : String(val);
