+++ cli/lib/commands/runtime.mjs
@@ -16,12 +16,8 @@
  */
 
 import { get, post } from "../api.mjs";
+import { truncate } from "../ui/ansi.mjs";
 
-function truncate(str, len) {
-  if (!str) return "";
-  if (str.length <= len) return str;
-  return str.slice(0, len - 3) + "...";
-}
 
 export async function showRuntimeToken() {
   const data = await get("/api/runtime/change-token");
