+++ web/lib/types/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Central re-export for shared web types. Pages import from
+ * `@/lib/types` and get everything; domain-specific modules live in
+ * sibling files so they can be imported individually too.
+ */
+
+export type { IdeaQuestion, IdeaWithScore } from "./ideas";
